+++ src/plugin/packer/__tests__/.gitignore
@@ -1,0 +1,2 @@
+fixtures/sample-plugin/plugin.zip
+fixtures/sample-plugin/*.ppk
