+++ src/plugin/packer/__tests__/bin.test.ts
