+++ src/plugin/packer/__tests__/fixtures/aaa.ppk
