+++ src/plugin/packer/__tests__/fixtures/contents.zip
