+++ src/plugin/packer/__tests__/fixtures/invalid-maxFileSize-contents.zip
