+++ src/plugin/packer/__tests__/fixtures/plugin-full-manifest/manifest.json
