+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-https-url/image/icon.png
