+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-https-url/manifest.json
