+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-maxFileSize/manifest.json
