+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-relative-path/image/icon.png
