+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-relative-path/manifest.json
