+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-url/image/icon.png
