+++ src/plugin/packer/__tests__/fixtures/plugin-invalid-url/manifest.json
