+++ src/plugin/packer/__tests__/fixtures/plugin-non-file-exists/manifest.json
