+++ src/plugin/packer/__tests__/fixtures/plugin.zip
