+++ src/plugin/packer/__tests__/fixtures/private.ppk
