+++ src/plugin/packer/__tests__/fixtures/sample-plugin/plugin-dir/image/icon.png
