+++ src/plugin/packer/__tests__/hex2a.test.ts
@@ -1,4 +1,4 @@
-import { hex2a } from "../src/hex2a";
+import { hex2a } from "../hex2a";
 
 describe("hex2a", () => {
   it("empty string", () => {
