--- src/plugin/packer/bin/eslint.config.mjs
@@ -1,4 +1,0 @@
-import presetsNodePrettier from "@cybozu/eslint-config/flat/presets/node-prettier.js";
-
-/** @type {import("eslint").Linter.Config[]} */
-export default [...presetsNodePrettier];
