--- src/plugin/packer/old.package.json
@@ -1,82 +1,0 @@
-{
-  "name": "@kintone/plugin-packer",
-  "description": "Package your kintone plugin with pure JavaScript",
-  "version": "8.1.3",
-  "author": {
-    "name": "Cybozu, Inc.",
-    "url": "https://cybozu.co.jp"
-  },
-  "engines": {
-    "node": ">=18"
-  },
-  "main": "dist/index.js",
-  "bin": {
-    "kintone-plugin-packer": "bin/cli.js"
-  },
-  "types": "dist/index.d.ts",
-  "files": [
-    "bin",
-    "dist",
-    "from-manifest.js"
-  ],
-  "scripts": {
-  },
-  "dependencies": {
-    "@kintone/plugin-manifest-validator": "^10.2.2",
-    "chokidar": "^3.6.0",
-    "debug": "^4.3.7",
-    "denodeify": "^1.2.1",
-    "meow": "^9.0.0",
-    "mkdirp": "^3.0.1",
-    "node-rsa": "~1.1.1",
-    "stream-buffers": "^3.0.3",
-    "yauzl": "^3.1.3",
-    "yazl": "^2.5.1"
-  },
-  "devDependencies": {
-    "@reduxjs/toolkit": "^2.3.0",
-    "@types/debug": "^4.1.12",
-    "@types/node-rsa": "^1.1.4",
-    "@types/stream-buffers": "^3.0.7",
-    "@types/yauzl": "^2.10.3",
-    "@types/yazl": "^2.4.5",
-    "ajv": "^8.17.1",
-    "array-flatten": "^3.0.0",
-    "assert": "^2.1.0",
-    "babel-plugin-replace-ts-export-assignment": "^0.0.2",
-    "browserify-zlib": "^0.2.0",
-    "buffer": "^6.0.3",
-    "constants-browserify": "^1.0.0",
-    "crypto-browserify": "^3.12.0",
-    "cssnano": "^7.0.6",
-    "execa": "^5.1.1",
-    "glob": "^10.4.5",
-    "globals": "^15.11.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "normalize.css": "^8.0.1",
-    "path-browserify": "^1.0.1",
-    "postcss": "^8.4.47",
-    "postcss-cli": "^11.0.0",
-    "process": "^0.11.10",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^3.1.0",
-    "rimraf": "^5.0.10",
-    "setimmediate": "^1.0.5",
-    "stream-browserify": "^3.0.0",
-    "util": "^0.12.5",
-    "webpack": "^5.95.0",
-    "webpack-cli": "5.1.4",
-    "webpack-dev-server": "^5.1.0"
-  },
-  "homepage": "https://github.com/kintone/js-sdk/tree/main/packages/plugin-packer",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/kintone/js-sdk.git",
-    "directory": "packages/plugin-packer"
-  },
-  "bugs": "https://github.com/kintone/js-sdk/issues",
-  "keywords": [
-    "kintone"
-  ],
-  "license": "MIT"
-}
